requirements.txt
arrow==1.0.3
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cssselect==1.1.0
date-guesser==2.1.4
decorator==4.4.2
Django==3.1.7
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.4
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-secret-manager==2.5.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
hvac==0.10.8
idna==2.10
jieba3k==0.35.1
joblib==1.0.1
libcst==0.3.19
loguru==0.5.3
lxml==4.6.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
newspaper3k==0.2.8
nltk==3.5
packaging==21.0
Pillow==8.2.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
sgmllib3k==1.0.0
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tinysegmenter==0.3
tld==0.12.5
tldextract==3.1.0
tqdm==4.59.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
validators==0.18.2
